# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book Inventory API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/libraries/{library_id}/enrichment/books/{library_book_id}": {
      "post": {
        "tags": [
          "enrichment"
        ],
        "summary": "Enrich Book",
        "operationId": "enrich_book_api_libraries__library_id__enrichment_books__library_book_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "library_book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryBookDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/jobs/{job_id}": {
      "post": {
        "tags": [
          "enrichment"
        ],
        "summary": "Process Job",
        "operationId": "process_job_api_libraries__library_id__enrichment_jobs__job_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryBookDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/books/{library_book_id}/candidate": {
      "get": {
        "tags": [
          "enrichment"
        ],
        "summary": "Get Candidate",
        "operationId": "get_candidate_api_libraries__library_id__enrichment_books__library_book_id__candidate_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "library_book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/books/{library_book_id}/candidate/apply": {
      "post": {
        "tags": [
          "enrichment"
        ],
        "summary": "Apply Candidate",
        "operationId": "apply_candidate_api_libraries__library_id__enrichment_books__library_book_id__candidate_apply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "library_book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryBookDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/books/{library_book_id}/candidate/reject": {
      "post": {
        "tags": [
          "enrichment"
        ],
        "summary": "Reject Candidate",
        "operationId": "reject_candidate_api_libraries__library_id__enrichment_books__library_book_id__candidate_reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "library_book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryBookDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/search": {
      "get": {
        "tags": [
          "enrichment"
        ],
        "summary": "Search Books Endpoint",
        "operationId": "search_books_endpoint_api_libraries__library_id__enrichment_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "search_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "auto",
              "title": "Search Type"
            }
          },
          {
            "name": "max_results",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Max Results"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Search Books Endpoint Api Libraries  Library Id  Enrichment Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/libraries/{library_id}/enrichment/preview/{identifier}": {
      "get": {
        "tags": [
          "enrichment"
        ],
        "summary": "Get Metadata Preview",
        "operationId": "get_metadata_preview_api_libraries__library_id__enrichment_preview__identifier__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Library Id"
            }
          },
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Metadata Preview Api Libraries  Library Id  Enrichment Preview  Identifier  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookV2Read": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "authors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authors"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "publish_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Date"
          },
          "subjects": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjects"
          },
          "language": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url",
            "description": "External cover URL from metadata provider"
          },
          "metadata_status": {
            "type": "string",
            "title": "Metadata Status",
            "description": "pending|awaiting_review|complete|failed",
            "default": "pending"
          },
          "metadata_candidate": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Candidate",
            "description": "Suggested metadata values awaiting review"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "BookV2Read",
        "description": "Schema for reading book data."
      },
...
      "CandidateApplyRequest": {
        "properties": {
          "fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fields",
            "description": "Fields to apply from the candidate data. Empty list applies all."
          }
        },
        "type": "object",
        "title": "CandidateApplyRequest"
      },
      "CandidateResponse": {
        "properties": {
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id"
          },
          "metadata_candidate": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Candidate"
          },
          "metadata_status": {
            "type": "string",
            "title": "Metadata Status"
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "metadata_candidate",
          "metadata_status"
        ],
        "title": "CandidateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryBookDetail": {
        "properties": {
          "book": {
            "$ref": "#/components/schemas/BookV2Read"
          },
          "library_book": {
            "$ref": "#/components/schemas/LibraryBookRead"
          },
          "personal_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBookDataRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "book",
          "library_book"
        ],
        "title": "LibraryBookDetail"
      },
...
      "LibraryBookRead": {
        "properties": {
          "condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition",
            "description": "Physical condition: New, Good, Fair, Poor"
          },
          "physical_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Location",
            "description": "Shelf, box, or storage location identifier"
          },
          "book_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Type",
            "description": "Format type: paperback, hardcover, ebook, pdf, audiobook"
          },
          "series": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series",
            "description": "Series name this physical copy belongs to"
          },
          "acquisition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acquisition Date",
            "description": "Date this copy was acquired by the library"
          },
          "loan_status": {
            "type": "string",
            "title": "Loan Status",
            "description": "available|checked_out",
            "default": "available"
          },
          "checked_out_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Out To",
            "description": "User ID of borrower if checked out"
          },
          "checked_out_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Out At",
            "description": "When the book was checked out"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "Due date for the current loan, if any"
          },
          "cover_image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Path",
            "description": "Local file path for uploaded cover image"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id"
          },
          "library_id": {
            "type": "string",
            "format": "uuid",
            "title": "Library Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "book_id",
          "library_id",
          "created_at",
          "updated_at"
        ],
        "title": "LibraryBookRead",
        "description": "Schema for reading library book data."
      },
...
      "UserBookDataRead": {
        "properties": {
          "reading_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Status",
            "description": "to-read, reading, completed, abandoned"
          },
          "progress_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Pages",
            "description": "Current page number"
          },
          "progress_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Percent",
            "description": "Reading progress as percentage (0-100)"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "Date user started reading this book"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Date user completed reading this book"
          },
          "personal_rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Rating",
            "description": "User's rating (1-5 stars)"
          },
          "personal_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Notes",
            "description": "User's private notes about this book"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite",
            "description": "Whether user marked this as favorite",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "library_id": {
            "type": "string",
            "format": "uuid",
            "title": "Library Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "book_id",
          "user_id",
          "library_id",
          "created_at",
          "updated_at"
        ],
        "title": "UserBookDataRead",
        "description": "Schema for reading user book data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}